ArmInteractionType fix
- removed `PonderArmInteractionPointType`
- refactored Type and Point clazz to `CentralKitchenArmInteractionTypes`, similar to Create.
