--- src/main/java/plus/dragons/createcentralkitchen/content/logistics/block/mechanicalArm/BlazeStovePoint.java
@@ -1,74 +1,0 @@
-package plus.dragons.createcentralkitchen.content.logistics.block.mechanicalArm;
-
-import com.simibubi.create.content.kinetics.mechanicalArm.AllArmInteractionPointTypes;
-import com.simibubi.create.content.kinetics.mechanicalArm.ArmInteractionPoint;
-import com.simibubi.create.content.kinetics.mechanicalArm.ArmInteractionPointType;
-import com.simibubi.create.foundation.ponder.PonderRegistry;
-import com.simibubi.create.infrastructure.ponder.AllPonderTags;
-import net.minecraft.core.BlockPos;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.Containers;
-import net.minecraft.world.InteractionResultHolder;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.ItemLike;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.phys.Vec3;
-import org.jetbrains.annotations.Nullable;
-import plus.dragons.createcentralkitchen.content.contraptions.blazeStove.BlazeStoveBlock;
-import plus.dragons.createcentralkitchen.content.contraptions.blazeStove.BlazeStoveBlockEntity;
-import plus.dragons.createcentralkitchen.entry.block.FDBlockEntries;
-import plus.dragons.createcentralkitchen.foundation.ponder.PonderArmInteractionPointType;
-
-import java.util.function.Consumer;
-
-public class BlazeStovePoint extends AllArmInteractionPointTypes.DepositOnlyArmInteractionPoint {
-
-    public BlazeStovePoint(ArmInteractionPointType type, Level level, BlockPos pos, BlockState state) {
-        super(type, level, pos, state);
-    }
-    
-    @Override
-    protected Vec3 getInteractionPositionVector() {
-        return Vec3.upFromBottomCenterOf(pos, 1);
-    }
-    
-    @Override
-    public ItemStack insert(ItemStack stack, boolean simulate) {
-        ItemStack input = stack.copy();
-        InteractionResultHolder<ItemStack> res = BlazeStoveBlock.tryInsert(level, pos, input, false, false, simulate);
-        ItemStack remainder = res.getObject();
-        if (input.isEmpty()) {
-            return remainder;
-        } else {
-            if (!simulate)
-                Containers.dropItemStack(level, pos.getX(), pos.getY(), pos.getZ(), remainder);
-            return input;
-        }
-    }
-    
-    public static class Type extends PonderArmInteractionPointType {
-        
-        public Type(ResourceLocation id) {
-            super(id);
-        }
-        
-        @Override
-        public boolean canCreatePoint(Level level, BlockPos pos, BlockState state) {
-            return level.getBlockEntity(pos) instanceof BlazeStoveBlockEntity;
-        }
-        
-        @Nullable
-        @Override
-        public ArmInteractionPoint createPoint(Level level, BlockPos pos, BlockState state) {
-            return new BlazeStovePoint(this, level, pos, state);
-        }
-    
-        @Override
-        public void addToPonderTag(Consumer<ItemLike> consumer) {
-            PonderRegistry.TAGS.forTag(AllPonderTags.ARM_TARGETS).add(FDBlockEntries.BLAZE_STOVE.get());
-        }
-        
-    }
-    
-}
