--- src/main/java/plus/dragons/createcentralkitchen/content/logistics/block/mechanicalArm/KettlePoint.java
@@ -1,216 +1,0 @@
-package plus.dragons.createcentralkitchen.content.logistics.block.mechanicalArm;
-
-import com.simibubi.create.content.kinetics.mechanicalArm.ArmInteractionPoint;
-import com.simibubi.create.content.kinetics.mechanicalArm.ArmInteractionPointType;
-import com.simibubi.create.foundation.ponder.PonderRegistry;
-import com.simibubi.create.infrastructure.ponder.AllPonderTags;
-import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
-import net.minecraft.world.item.alchemy.PotionUtils;
-import net.minecraft.world.item.alchemy.Potions;
-import net.minecraft.world.level.ItemLike;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.entity.BlockEntity;
-import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.material.Fluids;
-import net.minecraft.world.phys.Vec3;
-import net.minecraftforge.common.util.LazyOptional;
-import net.minecraftforge.items.IItemHandler;
-import org.jetbrains.annotations.Nullable;
-import plus.dragons.createcentralkitchen.content.contraptions.blazeStove.BlazeStoveBlockEntity;
-import plus.dragons.createcentralkitchen.content.logistics.item.guide.brewing.BrewingGuide;
-import plus.dragons.createcentralkitchen.content.logistics.item.guide.brewing.BrewingGuideItem;
-import plus.dragons.createcentralkitchen.foundation.ponder.PonderArmInteractionPointType;
-import umpaz.farmersrespite.common.block.KettleBlock;
-import umpaz.farmersrespite.common.block.entity.KettleBlockEntity;
-import umpaz.farmersrespite.common.registry.FRItems;
-import vectorwing.farmersdelight.common.utility.ItemUtils;
-
-import java.util.function.Consumer;
-
-public class KettlePoint extends ArmInteractionPoint {
-    // TODO Need change. Kettle changed a lot.
-    public static final int INPUT_SLOT_COUNT = 2;
-    public static final int CONTAINER_SLOT = 3;
-    public static final int OUTPUT_SLOT = 4;
-    
-    public KettlePoint(ArmInteractionPointType type, Level level, BlockPos pos, BlockState state) {
-        super(type, level, pos, state);
-    }
-    
-    @Override
-    protected Vec3 getInteractionPositionVector() {
-        return Vec3.upFromBottomCenterOf(pos, .625);
-    }
-    
-    @SuppressWarnings({"NullableProblems", "ConstantConditions"})
-    @Nullable
-    @Override
-    protected IItemHandler getHandler() {
-        if (!cachedHandler.isPresent()) {
-            BlockEntity te = level.getBlockEntity(pos);
-            if (!(te instanceof KettleBlockEntity kettle))
-                return null;
-            cachedHandler = LazyOptional.of(kettle::getInventory);
-        }
-        return cachedHandler.orElse(null);
-    }
-    
-    @Override
-    public ItemStack insert(ItemStack stack, boolean simulate) {
-        if (!(level.getBlockEntity(pos.below()) instanceof BlazeStoveBlockEntity blazeStove))
-            return stack;
-    
-        ItemStack guideStack = blazeStove.getGuide();
-        if (!(guideStack.getItem() instanceof BrewingGuideItem))
-            return stack;
-    
-        BrewingGuide guide = BrewingGuide.of(guideStack);
-        if (guide.getResult().isEmpty())
-            return stack;
-        
-        IItemHandler inventory = getHandler();
-        if (inventory == null)
-            return stack;
-        
-        if (inventory.getStackInSlot(CONTAINER_SLOT).isEmpty() && guide.isContainer(stack))
-            return inventory.insertItem(CONTAINER_SLOT, stack, simulate);
-    
-        boolean[] neededSlots = new boolean[INPUT_SLOT_COUNT];
-        int neededSlotCount = 0;
-        for (int slot = 0; slot < INPUT_SLOT_COUNT; slot++) {
-            if (inventory.getStackInSlot(slot).isEmpty() &&
-                guide.needIngredient(slot) &&
-                guide.isIngredient(slot, stack))
-            {
-                neededSlots[slot] = true;
-                neededSlotCount++;
-            }
-        }
-        
-        if (neededSlotCount == 0) {
-            return insertWater(guide, inventory, stack, simulate);
-        }
-
-        ItemStack ret = stack.copy();
-        ret.shrink(neededSlotCount);
-        for (int slot = 0; slot < INPUT_SLOT_COUNT; slot++) {
-            if (neededSlots[slot]) {
-                ItemStack inserted = stack.copy();
-                inserted.setCount(1);
-                inventory.insertItem(slot, inserted, simulate);
-            }
-        }
-        return ret;
-    }
-    
-    private ItemStack insertWater(BrewingGuide guide, IItemHandler inventory, ItemStack stack, boolean simulate) {
-        BlockState state = level.getBlockState(pos);
-        BlockEntity be = level.getBlockEntity(pos);
-        if(be instanceof KettleBlockEntity kettle){
-            var tank = kettle.getFluidTank();
-            if(tank.isEmpty() || tank.getFluid().getFluid().isSame(Fluids.WATER)){
-                var waterLevel = tank.getFluidAmount();
-                if (waterLevel==tank.getCapacity())
-                    return stack;
-
-                ItemStack remainder = stack.copy();
-                ItemStack container = ItemStack.EMPTY;
-
-                if (stack.is(Items.WATER_BUCKET)) {
-                    if(waterLevel+100>tank.getCapacity())
-                        return stack;
-                    container = Items.BUCKET.getDefaultInstance();
-
-                } else if (stack.is(Items.POTION) && PotionUtils.getPotion(stack) == Potions.WATER) {
-                    int canInsert = (tank.getCapacity() - waterLevel) / 250;
-                    if(canInsert<1)
-                        return stack;
-                    container = new ItemStack(Items.GLASS_BOTTLE, canInsert);
-                    remainder.shrink(canInsert);
-                }
-
-                if (!simulate) {
-                    if (guide.isContainer(container))
-                        container = inventory.insertItem(CONTAINER_SLOT, container, false);
-
-                    if (!container.isEmpty()) {
-                        Direction direction = state.getValue(KettleBlock.FACING).getCounterClockWise();
-                        double x = pos.getX() + 0.5D + (double) direction.getStepX() * 0.25D;
-                        double y = pos.getY() + 0.7D;
-                        double z = pos.getZ() + 0.5D + (double) direction.getStepZ() * 0.25D;
-                        ItemUtils.spawnItemEntity(this.level, container, x, y, z, direction.getStepX() * 0.08D, 0.25D, direction.getStepZ() * 0.08D);
-                    }
-
-                    tank.getFluid().setAmount(waterLevel);
-                }
-                return remainder;
-
-            }
-
-        }
-        return stack;
-    }
-    
-    @Override
-    public ItemStack extract(int slot, int amount, boolean simulate) {
-        if (slot == OUTPUT_SLOT)
-            return super.extract(slot, amount, simulate);
-    
-        if (!(level.getBlockEntity(pos.below()) instanceof BlazeStoveBlockEntity blazeStove))
-            return ItemStack.EMPTY;
-    
-        ItemStack guideStack = blazeStove.getGuide();
-        if (!(guideStack.getItem() instanceof BrewingGuideItem))
-            return ItemStack.EMPTY;
-    
-        BrewingGuide guide = BrewingGuide.of(guideStack);
-        if (guide.getResult().isEmpty())
-            return ItemStack.EMPTY;
-    
-        IItemHandler inventory = getHandler();
-        if (inventory == null)
-            return ItemStack.EMPTY;
-    
-        if (slot < INPUT_SLOT_COUNT) {
-            ItemStack ingredient = inventory.getStackInSlot(slot);
-            if (!ingredient.isEmpty() && !guide.isIngredient(slot, ingredient))
-                return inventory.extractItem(slot, amount, simulate);
-        } else if(slot == CONTAINER_SLOT) {
-            ItemStack container = inventory.getStackInSlot(slot);
-            if (!container.isEmpty() && !guide.isContainer(container))
-                return inventory.extractItem(slot, amount, simulate);
-        }
-        
-        return ItemStack.EMPTY;
-    }
-    
-    public static class Type extends PonderArmInteractionPointType {
-        
-        public Type(ResourceLocation id) {
-            super(id);
-        }
-        
-        @Override
-        public boolean canCreatePoint(Level level, BlockPos pos, BlockState state) {
-            return level.getBlockEntity(pos) instanceof KettleBlockEntity &&
-                level.getBlockEntity(pos.below()) instanceof BlazeStoveBlockEntity;
-        }
-        
-        @Nullable
-        @Override
-        public ArmInteractionPoint createPoint(Level level, BlockPos pos, BlockState state) {
-            return new KettlePoint(this, level, pos, state);
-        }
-    
-        @Override
-        public void addToPonderTag(Consumer<ItemLike> consumer) {
-            PonderRegistry.TAGS.forTag(AllPonderTags.ARM_TARGETS).add(FRItems.KETTLE.get());
-        }
-        
-    }
-    
-}
